- do not "migrate" already symlinked /usr/share/file
Changed files:
    file.spec -> 1.181

+++ file.spec
@@ -333,7 +333,7 @@ rm -rf $RPM_BUILD_ROOT
 
 %pretrans
 # it used to be directory
-if [ -d %{_datadir}/file ]; then
+if [ -d %{_datadir}/file -a ! -L %{_datadir}/file ]; then
 	mv -b %{_datadir}/file{,.dir}
 	ln -s misc %{_datadir}/file
 %banner -e %{name} <<EOF
